+++ mrpt_examples_cpp/graphslam_example/main.cpp
@@ -155,9 +155,9 @@ struct ExampleDemoGraphSLAM
     for (TNodeID j = 0; j < N_VERTEX; j++)
     {
       // Use evenly distributed nodes:
-      static double ang = 2 * M_PI / N_VERTEX;
+      static double ang = 2 * M_PI / static_cast<double>(N_VERTEX);
       const double R = NODES_XY_MAX + 2 * (j % 2 ? 1 : -1);
-      CPose2D p(R * cos(ang * j), R * sin(ang * j), ang);
+      CPose2D p(R * cos(ang * static_cast<double>(j)), R * sin(ang * static_cast<double>(j)), ang);
 
       // Save real pose:
       real_node_poses[j] = p;
@@ -264,9 +264,12 @@ struct ExampleDemoGraphSLAM
     log_sq_err_evolution.clear();
 
     std::cout << "Global graph RMS error / edge = "
-              << std::sqrt(graph.getGlobalSquareError(false) / graph.edgeCount()) << "\n";
+              << std::sqrt(
+                     graph.getGlobalSquareError(false) / static_cast<double>(graph.edgeCount()))
+              << "\n";
     std::cout << "Global graph RMS error / edge = "
-              << std::sqrt(graph.getGlobalSquareError(true) / graph.edgeCount())
+              << std::sqrt(
+                     graph.getGlobalSquareError(true) / static_cast<double>(graph.edgeCount()))
               << " (ignoring information matrices)."
               << "\n";
 
@@ -278,9 +281,12 @@ struct ExampleDemoGraphSLAM
         params, &my_levmarq_feedback<my_graph_t>);
 
     std::cout << "Global graph RMS error / edge = "
-              << std::sqrt(graph.getGlobalSquareError(false) / graph.edgeCount()) << "\n";
+              << std::sqrt(
+                     graph.getGlobalSquareError(false) / static_cast<double>(graph.edgeCount()))
+              << "\n";
     std::cout << "Global graph RMS error / edge = "
-              << std::sqrt(graph.getGlobalSquareError(true) / graph.edgeCount())
+              << std::sqrt(
+                     graph.getGlobalSquareError(true) / static_cast<double>(graph.edgeCount()))
               << " (ignoring information matrices)."
               << "\n";
